+++ django_test_app/server/apps/negotiations/views.py
@@ -9,7 +9,7 @@ from django.http import HttpRequest, HttpResponse
 from typing_extensions import override
 
 from dmr import Body, Controller, ResponseSpec, validate
-from dmr.compiled import request_accepts
+from dmr.compiled import accepted_header
 from dmr.exceptions import (
     DataRenderingError,
     RequestSerializationError,
@@ -123,7 +123,7 @@ class ContentNegotiationController(Controller[PydanticSerializer]):
             ContentType.xml: _RequestModel,
         }),
     ]:
-        if request_accepts(self.request, ContentType.json):
+        if accepted_header(self.request.headers, ContentType.json):
             return [
                 parsed_body.payment_method_id,
                 parsed_body.payment_amount,
@@ -143,7 +143,7 @@ class ContentNegotiationController(Controller[PydanticSerializer]):
         ),
     )
     def put(self, parsed_body: Body[_RequestModel]) -> HttpResponse:
-        if request_accepts(self.request, ContentType.json):
+        if accepted_header(self.request.headers, ContentType.json):
             return self.to_response(
                 [
                     parsed_body.payment_method_id,