+++ tests/test_unit/test_negotiation/test_global_configuration.py
@@ -11,7 +11,7 @@ from inline_snapshot import snapshot
 from typing_extensions import TypedDict, override
 
 from dmr import Body, Controller, ResponseSpec, modify, validate
-from dmr.compiled import request_accepts
+from dmr.compiled import accepted_header
 from dmr.errors import ErrorType
 from dmr.negotiation import ContentType, conditional_type, request_parser
 from dmr.parsers import JsonParser
@@ -404,7 +404,7 @@ def test_conditional_content_type(
                 ContentType.xml: dict[str, str],
             }),
         ]:
-            if request_accepts(self.request, ContentType.json):
+            if accepted_header(self.request.headers, ContentType.json):
                 return parsed_body.root['key']
             return parsed_body.root
 
@@ -424,7 +424,7 @@ def test_conditional_content_type(
             self,
             parsed_body: Body[Annotated[_RequestModel, 'other comment']],
         ) -> HttpResponse:
-            if request_accepts(self.request, ContentType.json):
+            if accepted_header(self.request.headers, ContentType.json):
                 return self.to_response(
                     parsed_body.root['key'],
                     status_code=HTTPStatus.CREATED,
@@ -503,7 +503,7 @@ def test_wrong_conditional_content_type(
             }),
         ]:
             # ERROR! Type to content logic is reversed:
-            if request_accepts(self.request, ContentType.json):
+            if accepted_header(self.request.headers, ContentType.json):
                 return parsed_body.root['key']
             return parsed_body.root
 
@@ -522,7 +522,7 @@ def test_wrong_conditional_content_type(
         )
         def put(self, parsed_body: Body[_RequestModel]) -> HttpResponse:
             # ERROR! Type to content logic is reversed:
-            if request_accepts(self.request, ContentType.json):
+            if accepted_header(self.request.headers, ContentType.json):
                 return self.to_response(
                     parsed_body.root['key'],
                     status_code=HTTPStatus.CREATED,
@@ -838,7 +838,7 @@ def test_conditional_error_model(
                 loc=loc,
                 error_type=error_type,
             )
-            if request_accepts(self.request, ContentType.json):
+            if accepted_header(self.request.headers, ContentType.json):
                 return {
                     'json_errors': [
                         {'reason': f'{detail["msg"]}: {detail["loc"]}'}
@@ -925,7 +925,7 @@ def test_conditional_error_model_wrong(
                 error_type=error_type,
             )
             # NOTE: we change the formats to trigger the validation:
-            if request_accepts(self.request, ContentType.xml):
+            if accepted_header(self.request.headers, ContentType.xml):
                 return {
                     'json_errors': [
                         {'reason': f'{detail["msg"]}: {detail["loc"]}'}